Conflicts:
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/api/NotificationDao.java
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/identification/NaturalPersonPrimaryRnuRowMapper.java
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/NdflPersonDaoImpl.java
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/NotificationDaoImpl.java
	dao/src/test/java/com/aplana/sbrf/taxaccounting/dao/impl/NdflPersonDaoTest.java
	database/distr/v1.0/create_procedures.sql
	database/distr/v1.0/templates/data/ndfl/consolidated_rnu_ndfl/v2016/script.groovy
	database/distr/v1.0/templates/data/ndfl/primary_rnu_ndfl/v2016/script.groovy
	database/distr/v1.0/templates/data/ndfl/report_2ndfl_1/v2016/script.groovy
	database/distr/v1.0/templates/data/ndfl/report_6ndfl/v2016/script.groovy
	database/distr/v1.0_next/templates/data/ndfl/primary_rnu_ndfl/v2016/report_person_db.jrxml
	database/distr/v1.0_next/templates/data/ndfl/primary_rnu_ndfl/v2016/rnu_ndfl_light.xsd
	database/distr/v1.0_next/templates/data/ndfl/report_2ndfl_1/v2016/report.jrxml
	database/distr/v1.0_next/templates/data/ndfl/report_2ndfl_1/v2016/script.groovy
	database/distr/v1.0_next/templates/data/refbook/fias/script.groovy
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationdata/client/DeclarationDataPresenter.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationdata/client/DeclarationDataView.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationlist/client/DeclarationListModule.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationlist/client/DeclarationListView.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationtemplate/client/editform/EditFormPresenter.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/formdata/client/sources/SourcesView.ui.xml
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/sources/server/GetCurrentAssignsHandler.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/taxformnomination/server/DeleteFormSoursesHandler.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/taxformnomination/server/GetTaxFormTypesHandler.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/mvc/CacheController.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/widget/signin/client/SignInPresenter.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/widget/signin/client/SignInView.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/widget/signin/client/SignInView.ui.xml
	gwtapp/src/main/webapp/WEB-INF/security.xml
	gwtapp/src/main/webapp/WEB-INF/web.xml
	gwtapp/src/main/webapp/resources/help_conf.pdf
	gwtapp/src/test/config/jetty-realm.properties
	script-test/src/test/java/com/aplana/sbrf/taxaccounting/form_template/ndfl/primary_rnu_ndfl/v2016/PrimaryRnuNdflScriptCheckTest.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataScriptingServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/FormTemplateImpexServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/RefBookScriptingServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/TAAbstractScriptingServiceImpl.java
	service/src/test/java/com/aplana/sbrf/taxaccounting/service/impl/PrintingServiceImplTest.java
	service/src/test/java/com/aplana/sbrf/taxaccounting/service/impl/ScriptExecutionServiceTest.java
	src/main/resources/ddl/create_procedures.sql
	src/main/resources/dml/refbook.sql
	src/main/resources/form_template/ndfl/report_2ndfl_1/v2016/script.groovy
	src/main/resources/form_template/ndfl/report_6ndfl/v2016/script.groovy
	src/main/resources/refbook/declaration_type/script.groovy

+++ script-test/src/test/java/com/aplana/sbrf/taxaccounting/form_template/ndfl/consolidated_rnu_ndfl/v2016/ConsolidatedRnuNdflScriptTest.java
@@ -1,20 +1,6 @@
 package com.aplana.sbrf.taxaccounting.form_template.ndfl.consolidated_rnu_ndfl.v2016;
 
-import com.aplana.sbrf.taxaccounting.model.DeclarationData;
-import com.aplana.sbrf.taxaccounting.model.DeclarationFormKind;
-import com.aplana.sbrf.taxaccounting.model.DeclarationTemplate;
-import com.aplana.sbrf.taxaccounting.model.Department;
-import com.aplana.sbrf.taxaccounting.model.DepartmentReportPeriod;
-import com.aplana.sbrf.taxaccounting.model.FormDataEvent;
-import com.aplana.sbrf.taxaccounting.model.FormSources;
-import com.aplana.sbrf.taxaccounting.model.PagingParams;
-import com.aplana.sbrf.taxaccounting.model.PagingResult;
-import com.aplana.sbrf.taxaccounting.model.Relation;
-import com.aplana.sbrf.taxaccounting.model.ReportPeriod;
-import com.aplana.sbrf.taxaccounting.model.State;
-import com.aplana.sbrf.taxaccounting.model.TAUserInfo;
-import com.aplana.sbrf.taxaccounting.model.TaxPeriod;
-import com.aplana.sbrf.taxaccounting.model.TaxType;
+import com.aplana.sbrf.taxaccounting.model.*;
 import com.aplana.sbrf.taxaccounting.model.log.Logger;
 import com.aplana.sbrf.taxaccounting.model.ndfl.NdflPerson;
 import com.aplana.sbrf.taxaccounting.model.ndfl.NdflPersonDeduction;
@@ -37,21 +23,11 @@ import org.mockito.stubbing.Answer;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Random;
+import java.util.*;
 
 import static com.aplana.sbrf.taxaccounting.util.TestUtils.toDate;
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyBoolean;
-import static org.mockito.Matchers.anyInt;
-import static org.mockito.Matchers.anyList;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Matchers.eq;
+import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.when;
 
 /**
@@ -285,7 +261,7 @@ public class ConsolidatedRnuNdflScriptTest extends DeclarationScriptTestBase {
 
     private NdflPersonIncome createNdflPersonIncomes(BigDecimal row, String operationId) {
         NdflPersonIncome personIncome = new NdflPersonIncome();
-        personIncome.setRowNum(row);
+        personIncome.setRowNum(BigDecimal.valueOf(row));
         personIncome.setOperationId("11111");
         personIncome.setOktmo("oktmo111");
         personIncome.setKpp("kpp111");
@@ -298,7 +274,7 @@ public class ConsolidatedRnuNdflScriptTest extends DeclarationScriptTestBase {
 
     private NdflPersonDeduction createNdflPersonDeduction(BigDecimal row) {
         NdflPersonDeduction personDeduction = new NdflPersonDeduction();
-        personDeduction.setRowNum(row);
+        personDeduction.setRowNum(BigDecimal.valueOf(row));
         personDeduction.setOperationId("11111");
         personDeduction.setTypeCode("001");
 
@@ -323,7 +299,7 @@ public class ConsolidatedRnuNdflScriptTest extends DeclarationScriptTestBase {
 
     private NdflPersonPrepayment createNdflPersonPrepayment(BigDecimal row) {
         NdflPersonPrepayment personPrepayment = new NdflPersonPrepayment();
-        personPrepayment.setRowNum(row);
+        personPrepayment.setRowNum(BigDecimal.valueOf(row));
         personPrepayment.setOperationId("11111");
         personPrepayment.setSumm(new BigDecimal(1999999)); //по xsd это поле xs:integer
         personPrepayment.setNotifNum("123-456-000");
